🔨 Refactor Navigation to use React Hooks, TypeScript
Applied coding conventions to the Sandbox Workspace Navigation component.

Removed unused SVGs and updated references in UserMenu and Dashboard Sidebar to use icons from the common lib instead.

Tested locally. No observable changes in behavior.
