--- packages/app/src/app/pages/Sandbox/Editor/Navigation/ConfigurationIcon.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-import IconBase from 'react-icons/lib/IconBase';
-
-export default () => (
-  <IconBase
-    width="1em"
-    height="1em"
-    viewBox="0 0 196 196"
-    version="1.1"
-    xmlnsXlink="http://www.w3.org/1999/xlink"
-  >
-    <g fill="currentColor" fillRule="nonzero" stroke="none" strokeWidth={1}>
-      <path d="M181.075 146.8c.22-1.754.384-3.509.384-5.373 0-1.864-.164-3.618-.384-5.373l11.568-9.046c1.042-.822 1.316-2.302.658-3.509l-10.965-18.97c-.658-1.206-2.138-1.645-3.344-1.206l-13.652 5.482c-2.85-2.193-5.92-4.002-9.266-5.373l-2.083-15.329c-.164-1.316-1.316-3.103-2.686-3.103h-21.93c-1.37 0-2.522 1.787-2.686 3.103l-2.083 14.929c-3.344 1.371-6.415 3.435-9.266 5.573l-13.65-5.382c-1.262-.493-2.687.05-3.345 1.256L87.38 123.474a2.718 2.718 0 0 0 .658 3.521l11.568 9.053c-.219 1.754-.384 3.567-.384 5.376 0 1.809.164 3.62.384 5.374l-11.568 9.047c-1.042.822-1.316 2.303-.658 3.509l10.965 18.97c.658 1.206 2.138 1.645 3.344 1.206l13.651-5.482c2.851 2.193 5.921 4.002 9.266 5.373l2.083 14.403c.164 1.316 1.316 2.176 2.686 2.176h21.93c1.37 0 2.522-.86 2.686-2.176l2.083-14.466c3.344-1.371 6.415-3.203 9.266-5.341l13.652 5.498c1.261.493 2.686.008 3.344-1.198l10.965-18.966c.658-1.206.384-2.685-.658-3.507l-11.568-9.044zm-40.735 13.816c-10.58 0-19.189-8.608-19.189-19.189s8.608-19.189 19.19-19.189c10.58 0 19.188 8.608 19.188 19.189s-8.608 19.189-19.189 19.189z" />
-      <path d="M78.857 158.115l10.519-9.896c-.269-2.151-.471-4.472-.471-6.69s.202-4.488.471-6.639l-9.652-9.187a3.331 3.331 0 0 1-.807-4.315l13.445-23.266a3.281 3.281 0 0 1 4.101-1.482l13.557 4.764c3.496-2.622 7.26-4.908 11.361-6.589l1.204-12.466c.202-1.614 1.614-2.349 3.294-2.349h19.064V52.039L91.649 0H20.73C10.979 0 3.089 7.403 3.089 17.154L3 159.631c0 9.751 7.89 18.368 17.641 18.368h65.604l-7.363-13.664c-.807-1.478-1.302-5.212-.025-6.22zm4.086-145.968L131.541 63H82.943V12.147z" />
-    </g>
-  </IconBase>
-);
