--- packages/app/src/app/pages/Sandbox/Editor/Navigation/FileIcon.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-import IconBase from 'react-icons/lib/IconBase';
-
-export default () => (
-  <IconBase
-    width="1em"
-    height="1em"
-    viewBox="0 0 40 40"
-    version="1.1"
-    xmlnsXlink="http://www.w3.org/1999/xlink"
-  >
-    <path
-      d="M8 0C5.8 0 4.02 1.8 4.02 4L4 36c0 2.2 1.78 4 3.98 4H32c2.2 0 4-1.8 4-4V12L24 0H8zm14 14V3l11 11H22z"
-      fill="currentColor"
-      fillRule="nonzero"
-    />
-  </IconBase>
-);
