--- packages/app/src/app/pages/Sandbox/Editor/Navigation/InfoIcon.js
@@ -1,23 +1,0 @@
-import React from 'react';
-
-import IconBase from 'react-icons/lib/IconBase';
-
-export default ({ width, height }) => (
-  <IconBase
-    width={width || '1em'}
-    height={height || '1em'}
-    viewBox="0 0 241 241"
-    version="1.1"
-    xmlnsXlink="http://www.w3.org/1999/xlink"
-  >
-    <g
-      transform="translate(15.000000, 0.000000)"
-      fill="currentColor"
-      fillRule="nonzero"
-      stroke="none"
-      strokeWidth={1}
-    >
-      <path d="M100.949 112.375L202.557 55.18 110.02 1.998a9.01 9.01 0 0 0-8.975 0L7.26 55.897l93.689 56.478zM91 126.96L1 71.688v106.971c0 3.224-.672 6.803 2.122 8.41L91 239.507V126.96zm20 .607V240.8l95.012-54.334c2.794-1.607 4.988-4.584 4.988-7.808V70.251l-100 57.316z" />
-    </g>
-  </IconBase>
-);
