--- packages/app/src/app/pages/Sandbox/Editor/Navigation/RocketIcon.js
@@ -1,24 +1,0 @@
-import React from 'react';
-
-import IconBase from 'react-icons/lib/IconBase';
-
-export default () => (
-  <IconBase
-    width="1em"
-    height="1em"
-    viewBox="0 0 255 255"
-    version="1.1"
-    xmlnsXlink="http://www.w3.org/1999/xlink"
-  >
-    <g
-      transform="translate(31.000000, 0.000000)"
-      fill="currentColor"
-      fillRule="nonzero"
-      stroke="none"
-      strokeWidth={1}
-    >
-      <path d="M182.381 1.852C166.135-7.389 110.913 26.68 80.923 78.595c-3.038 5.259-6.922 9.98-8.922 15.488v-.002c-7 11.428-25.14 32.879-38.422 38.225-15.115 5.174-28.14 28.777-31.988 37.685-1.544 3.573-4.144 8.316 6.928 12.505 6.909 2.614 10.774 4.272 13.021-.261 13.239-22.208 31.572-20.896 31.572-20.896l-4.215 12.205c-.882 1.526-.032 3.59 1.714 4.599l43.672 25.229c1.747 1.009 3.957.714 4.839-.813l9-9.183s10.02 15.414-2.894 37.813c-2.858 4.174.388 6.732 6.043 11.484 9.063 7.616 11.789 2.96 14.153-.132 5.894-7.709 19.862-30.773 16.995-46.489-1.821-14.202 6.503-41.637 13.654-52.794l-.045.048a240.289 240.289 0 0 0 9.878-15.667c29.991-51.915 32.702-116.556 16.475-125.787zm-34.335 93.989c-4.935 8.417-15.759 11.24-24.176 6.305-8.417-4.935-11.24-15.759-6.305-24.176 4.935-8.417 15.759-11.24 24.176-6.305 8.417 4.935 11.239 15.759 6.305 24.176z" />
-      <path d="M45.577 189.551c-7.526 14.527-19.455 59.208-10.18 64.646 9.275 5.438 41.145-27.171 50.361-40.171.784-1.106 3.094-3.547-1.866-6.455-6.901-4.046-16.953 8.897-24.621 4.402-7.321-4.292-2.337-20.207-8.592-23.874-2.804-1.644-4.492.275-5.102 1.452z" />
-    </g>
-  </IconBase>
-);
