+++ packages/app/src/app/pages/Sandbox/Editor/index.js
@@ -11,7 +11,7 @@ import { Container } from './elements';
 import Workspace from './Workspace';
 import Content from './Content';
 import Header from './Header';
-import Navigation from './Navigation';
+import { Navigation } from './Navigation';
 import getVSCodeTheme from './utils/get-vscode-theme';
 
 const STATUS_BAR_SIZE = 22;
