+++ packages/app/src/app/pages/common/UserMenu/Menu/index.js
@@ -14,7 +14,7 @@ import {
   dashboardUrl,
 } from '@codesandbox/common/lib/utils/url-generator';
 import PatronBadge from '@codesandbox/common/lib/utils/badges/PatronBadge';
-import InfoIcon from 'app/pages/Sandbox/Editor/Navigation/InfoIcon';
+import InfoIcon from '-!svg-react-loader!@codesandbox/common/lib/icons/sandbox.svg';
 import track from '@codesandbox/common/lib/utils/analytics';
 
 import { Container, Item, Icon, Separator } from './elements';
